feat: combine institutions GraphQL and REST apis

+++ apps/api/src/institutions/institutions.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, Logger, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  Logger,
+  NotFoundException,
+} from '@nestjs/common';
 import { Institution } from './models/institution.model';
 import { InjectRepository } from '@nestjs/typeorm';
 import { IsNull, Repository } from 'typeorm';
@@ -14,39 +19,66 @@ export class InstitutionsService {
     private repository: Repository<Institution>,
   ) {}
 
-  findAll(args: { user_id: string } | null = null): Promise<Institution[]> {
+  findAll(args: { user_id?: string }): Promise<Institution[]> {
+    this.logger.debug(`Getting all institutions for user ${args.user_id}`);
+
     // The institutions with null owner are common to everyone
-    if (args?.user_id) {
-      return this.repository.find({
-        where: [{ user_id: args.user_id }, { user_id: IsNull() }],
-      });
-    } else {
-      return this.repository.find();
-    }
+    return this.repository.find({
+      where: [{ user_id: args.user_id }, { user_id: IsNull() }],
+    });
   }
 
-  findOne(id: string): Promise<Institution> {
+  findOne(args: { id: string; user_id: string | null }): Promise<Institution> {
+    // We want the user institutions, OR the common to all users
+    return this.repository.findOne({
+      where: [
+        {
+          id: args.id,
+          user_id: args.user_id,
+        },
+        {
+          id: args.id,
+          user_id: IsNull(),
+        },
+      ],
+    });
+  }
+
+  findOneById(id: string): Promise<Institution> {
     return this.repository.findOne({
       where: {
         id: id,
-        // user_id: args.user_id,
       },
     });
   }
 
-  create(dto: CreateInstitutionRequest, user_id: string = null) {
+  create(args: { dto: CreateInstitutionRequest; user_id: string }) {
     return this.repository.save({
-      name: dto.name,
-      image_src: dto.image_src,
-      user_id: user_id,
+      name: args.dto.name,
+      image_src: args.dto.image_src,
+      user_id: args.user_id,
     });
   }
 
-  async update(id: string, dto: UpdateInstitutionRequest) {
-    const institutionToUpdate = await this.findOne(id);
+  async update(args: {
+    id: string;
+    dto: UpdateInstitutionRequest;
+    user_id: string;
+  }) {
+    const institutionToUpdate = await this.findOne({
+      id: args.id,
+      user_id: args.user_id,
+    });
+
     if (institutionToUpdate) {
-      institutionToUpdate.name = dto.name;
-      institutionToUpdate.image_src = dto.image_src;
+      if (institutionToUpdate.user_id === null) {
+        // You cannot update common institutions!
+        throw new ForbiddenException();
+      }
+
+      institutionToUpdate.name = args.dto.name ?? institutionToUpdate.name;
+      institutionToUpdate.image_src =
+        args.dto.image_src ?? institutionToUpdate.image_src;
 
       return await this.repository.save(institutionToUpdate);
     } else {
@@ -54,16 +86,24 @@ export class InstitutionsService {
     }
   }
 
-  async remove(id: string): Promise<Institution> {
-    const entityToDelete = await this.findOne(id);
+  async remove(args: { id: string; user_id: string }): Promise<Institution> {
+    const entityToDelete = await this.findOne({
+      id: args.id,
+      user_id: args.user_id,
+    });
     if (entityToDelete) {
+      if (entityToDelete.user_id === null) {
+        // You cannot delete common institutions!
+        throw new ForbiddenException();
+      }
+
       this.logger.debug(`Deleting institution id ${entityToDelete.id}`);
       const deleted = await this.repository.remove(entityToDelete);
 
       // Remove returns the object without the ID. So rehydrate the returned object with
       // the original id.
       // As alternative, it would be possible to use `softRemoved` which returns the ID
-      const result = { ...deleted, id: id };
+      const result = { ...deleted, id: args.id };
       return result;
     } else {
       throw new NotFoundException();
